Refactor: Camera 기능 확인
