+++ src/App.jsx
@@ -3,11 +3,12 @@ import {
   Routes,
   Route
 } from 'react-router-dom';
+import CameraPage from './pages/camera/CameraPage';
 
 function App() {
   return (
     <Routes>
-
+      <Route path='/camera' element={<CameraPage />} />
     </Routes>
   )
 }
