+++ vite.config.js
@@ -8,6 +8,8 @@ import path from 'path';
 export default defineConfig({
   plugins: [react()],
   server: {
+    host: '0.0.0.0',
+    port: 5173,
     https: {
       key: fs.readFileSync(path.resolve(__dirname, 'certs/localhost+2-key.pem')),
       cert: fs.readFileSync(path.resolve(__dirname, 'certs/localhost+2.pem'))
